requirements.txt
django
djangorestframework
aiogram
load_dotenv
pillow
celery
redis
django-celery-beat
cryptography